Fix RPS F6 autocreate
Autocreate of F6-02-03 (PTM2xx) with multibutton does not work. It is not possible to determine the desired EEP from the telegram status t21 and nu bits. So I changed the code in a way, that the desired EEP is available in the moment of createObjects(). I also implement a strict check of EEP and status bits according the SPEC.