Fix EEP F6-02-01 multibutton
The EEPs F6-02-01 and F6-02-02 are literally the same. The difference is only, which side is mounted up, A0/B0 or AI/B1. On the other hand F6-02-03, currently used for multibuttons, is already used for a Servodan switch without multibuttons. So I have renamed the EEP definitions. F6-02-01 has the multibuttons and F6-02-02 has it not. F6-02-03 is free to be used for the Servodan switch.